amazon_webscraping/requirements.txt
slackclient
pandas
bs4
webdriver_manager.chrome
selenium
python-dotenv
tabulate